fix int type

+++ entity/product.go
@@ -5,8 +5,8 @@ import "time"
 type Product struct {
 	ID        uint64
 	Name      string
-	Price     int64
-	Stock     int32
+	Price     int
+	Stock     int
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
